Some updates.
Indication of the ship's magazine was updated. Now then ship will have a maximum counts of bullets in his magazine then particles will be gona blue.

New random bonuses and bullets generation was added. The new function generate a inventory that collect random types of buffs, bonuses, and rules. Minimum and macximum counts of types and objects of each type was seted every function call.

New buff "Stream" was added. The buff increment maximum count of bullets in the ship's magazine.

Suped powered asteroids was added. A super powered asteroid is a asteroid that is containing any buff. A steroid can be powering and super powering in one time.

Textures for all bufs was added.

Added sublimation for asteroid's textures.